Commit 13, proyecto finalizado
Se han probado todos los controladores, cada metodo funciona bien.
Proyecto finalizado

+++ apiStarWarsProject/Services/JediService.cs
@@ -64,7 +64,7 @@ namespace apiStarWarsProject.Services
             
             jediDB.JediName = jediDTO.JediNameDTO;
             jediDB.JediRankId = jediDTO.JediRankIdDTO;
-            jediDB.ArmyDivisionJedi = (int)jediDTO.ArmyDivisionJediDTO;
+            jediDB.ArmyDivisionJedi = jediDTO.ArmyDivisionJediDTO;
 
             _context.jedis.Update(jediDB);
             await _context.SaveChangesAsync();
